+++ common/src/main/java/net/mehvahdjukaar/moonlight/core/integration/MapAtlasCompat.java
@@ -13,7 +13,10 @@ import org.jetbrains.annotations.Nullable;
 import pepjebs.mapatlases.MapAtlasesMod;
 import pepjebs.mapatlases.client.MapAtlasesClient;
 import pepjebs.mapatlases.item.MapAtlasItem;
-import pepjebs.mapatlases.map_collection.MapKey;
+import pepjebs.mapatlases.map_collection.MapCollection;
+import pepjebs.mapatlases.map_collection.MapSearchKey;
+import pepjebs.mapatlases.utils.MapDataHolder;
+import pepjebs.mapatlases.utils.Slice;
 
 public class MapAtlasCompat {
 
@@ -24,31 +27,15 @@ public class MapAtlasCompat {
     @Nullable
     public static MapItemSavedData getSavedDataFromAtlas(ItemStack atlas, Level level, Player player) {
         if(isAtlas(atlas.getItem())) {
-            var maps = MapAtlasItem.getMaps(atlas, level);
+            MapCollection maps = MapAtlasItem.getMaps(atlas, level);
             if (maps != null) {
-                var slice = MapAtlasItem.getSelectedSlice(atlas, level.dimension());
-                var key = MapKey.at(maps.getScale(), player, slice);
-                var select = maps.select(key);
+                Slice slice = MapAtlasItem.getSelectedSlice(atlas, level.dimension());
+                MapSearchKey key = MapSearchKey.at(maps.getScale(), player, slice);
+                MapDataHolder select = maps.select(key);
                 if (select != null) {
                     return select.data;
                 }
             }
-        }
-        return null;
-    }
-
-    @Nullable
-    public static Integer getMapIdFromAtlas(ItemStack atlas, Level level, Object data) {
-        try {
-            var maps = MapAtlasItem.getMaps(atlas, level);
-            if (maps != null) {
-                for (var e : maps.getAll()) {
-                    if (e.data == data) {
-                        return e.id;
-                    }
-                }
-            }
-        } catch (Exception ignored) {
         }
         return null;
     }